Change -[retainCount] back to size_t and cast, so we keep the old API.

+++ src/OFConstString.m
@@ -138,9 +138,9 @@ void *_OFConstStringClassReference;
 	return self;
 }
 
-- (int32_t)retainCount
+- (size_t)retainCount
 {
-	return INT32_MAX;
+	return SIZE_MAX;
 }
 
 - (void)release
